Fixing merge conflicts and splitting code to folders
