--- public/googleae0c887b93cdf191.html
@@ -1,1 +1,0 @@
-google-site-verification: googleae0c887b93cdf191.html
