+++ public/manifest.json
@@ -1,6 +1,6 @@
 {
-  "short_name": "React App",
-  "name": "Create React App Sample",
+  "short_name": "Winnipeg.js",
+  "name": "Winnipeg Javascript Community",
   "icons": [
     {
       "src": "favicon.ico",
