--- src/components/App.js
@@ -1,14 +1,0 @@
-import React from 'react'
-import Home from './Home'
-import Header from './Header'
-import Footer from './Footer'
-
-const App = () => {
-  return <div id="body_container">
-    <Header />
-    <Home />
-    <Footer />
-  </div>
-}
-
-export default App;
