+++ src/screens/Forum.jsx
@@ -1,0 +1,13 @@
+import React from 'react';
+import { Link } from 'react-router-dom';
+
+const Forum = () => {
+    return (
+        <div>
+            <h1>Forum</h1>
+            <Link to="/resources">Resources</Link>
+        </div>
+    );
+}
+
+export default Forum;
