+++ src/screens/Resources.jsx
@@ -1,0 +1,13 @@
+import React from 'react';
+import { Link } from 'react-router-dom';
+
+const Resources = () => {
+    return (
+        <div>
+            <h1>Resources</h1>
+            <Link to="/home">Home</Link>
+        </div>
+    );
+}
+
+export default Resources;
